refactor/valorhora-bigdecimal

+++ fivegears-backend/src/main/kotlin/com/fivegears/fivegears_backend/dto/UsuarioAlocadoDTO.kt
@@ -1,5 +1,7 @@
 package com.fivegears.fivegears_backend.dto
 
+import java.math.BigDecimal
+
 data class UsuarioAlocadoDTO(
     val id: Int,
     val nome: String,
@@ -7,7 +9,7 @@ data class UsuarioAlocadoDTO(
     val cargo: String?,
     val idCargo: Int?,
     val senioridade: String?,
-    val valorHora: Double,
+    val valorHora: BigDecimal,
     val horasDisponiveis: Int,
     val projetosAtivos: List<String>,
     val softSkills: Map<String, Int>,
